implement and test method averageGradeCourse

+++ 061224_lombok/src/main/java/org/example/Student.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.With;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Data
@@ -15,7 +16,7 @@ public class Student {
     private String id;
     private String name;
     private String address;
-    int grade;
+    BigDecimal grade;
 
     @With
     private String warning;
